New Relic Service Broker
Currently, if a user wants to bind a Cloud Foundry application to New
Relic on-premise, they must use a user-provided service.  This isn't
impossible, but it isn't as nice as the off-premise version of the
same workflow.  This project is a service broker for New Relic that
can be run on-premise.  It's very simple, in that it simply returns a
payload containing a static set of New Relic credentials that are
configured at startup.  When this service broker is connected to the
Cloud Foundry cloud controller, it enables users to use `create-
service` and `bind-service` as they normally would in the off-premise
scenario.